e3backend/requirements.txt
django
djangorestframework
django-cors-headers
pymysql
pillow
